src/test/java/com/leetcode/algorithm/easy/cousinsinbinarytree/SolutionTest.java
package com.leetcode.algorithm.easy.cousinsinbinarytree;

import static org.junit.jupiter.api.Assertions.assertEquals;

import com.leetcode.algorithm.common.TreeNode;
import org.junit.jupiter.api.Test;

public class SolutionTest {
  @Test
  public void testCase1() {
    Solution solution = new Solution();
    TreeNode root = new TreeNode(1);
    root.left = new TreeNode(2);
    root.left.left = new TreeNode(4);
    root.right = new TreeNode(3);

    assertEquals(false, solution.isCousins(root, 4, 3));
  }

  @Test
  public void testCase2() {
    Solution solution = new Solution();
    TreeNode root = new TreeNode(1);
    root.left = new TreeNode(2);
    root.left.right = new TreeNode(4);
    root.right = new TreeNode(3);
    root.right.right = new TreeNode(5);

    assertEquals(true, solution.isCousins(root, 5, 4));
  }

  @Test
  public void testCase3() {
    Solution solution = new Solution();
    TreeNode root = new TreeNode(1);
    root.left = new TreeNode(2);
    root.left.right = new TreeNode(4);
    root.right = new TreeNode(3);

    assertEquals(false, solution.isCousins(root, 2, 3));
  }
}

